Fix lquitas como no
